Lift common stuff and add more plots
