+++ extras/fonll_hq_treatment/kDISbThr.py
@@ -1,3 +1,4 @@
+"""Vary `kbThr` inside yadism."""
 import copy
 from typing import Tuple
 
@@ -5,90 +6,27 @@ import lhapdf
 import matplotlib.pyplot as plt
 import numpy as np
 import numpy.typing as npt
-from eko import interpolation
+from utils import build_q2_obs, observables_card, theory_card
 
 import yadism
-
-t = {
-    "ID": 400,
-    "PTO": 2,
-    "FNS": "FONLL-C",
-    "DAMP": 0,
-    "IC": 1,
-    "IB": 0,
-    "ModEv": "TRN",
-    "ModSV": "unvaried",
-    "XIR": 1.0,
-    "XIF": 1.0,
-    "fact_to_ren_scale_ratio": 1.0,
-    "NfFF": 5,
-    "MaxNfAs": 5,
-    "MaxNfPdf": 5,
-    "Q0": 1.65,
-    "alphas": 0.118,
-    "Qref": 91.2,
-    "nf0": None,
-    "nfref": None,
-    "QED": 0,
-    "alphaqed": 0.007496252,
-    "Qedref": 1.777,
-    "SxRes": 0,
-    "SxOrd": "LL",
-    "HQ": "POLE",
-    "mc": 1.51,
-    "Qmc": 1.51,
-    "kcThr": 1.0,
-    "mb": 4.92,
-    "Qmb": 4.92,
-    "kbThr": 1.0,
-    "mt": 172.5,
-    "Qmt": 172.5,
-    "ktThr": 1.0,
-    "CKM": "0.97428 0.22530 0.003470 0.22520 0.97345 0.041000 0.00862 0.04030 0.999152",
-    "MZ": 91.1876,
-    "MW": 80.398,
-    "GF": 1.1663787e-05,
-    "SIN2TW": 0.23126,
-    "TMC": 0,
-    "MP": 0.938,
-    "Comments": "NNPDF4.0 NNLO alphas=0.118",
-    "global_nx": 0,
-    "EScaleVar": 1,
-    "kDIScThr": 1.0,
-    "kDISbThr": 1.0,
-    "kDIStThr": 1.0,
-}
-
-o = {
-    "PolarizationDIS": 0.0,
-    "ProjectileDIS": "positron",
-    "PropagatorCorrection": 0.0,
-    "TargetDIS": "proton",
-    "interpolation_is_log": True,
-    "interpolation_polynomial_degree": 4,
-    "interpolation_xgrid": interpolation.make_lambert_grid(60),
-    "observables": {"XSHERANC": []},
-    "prDIS": "NC",
-    "NCPositivityCharge": None,
-}
 
 
 def compute(kDISbThrs: list, curobs: list) -> Tuple[npt.ArrayLike, list]:
     """Compute data for given observables"""
     q2s = np.array([esf["Q2"] for esf in curobs])
     # prepare data
-    oo = copy.deepcopy(o)
+    oo = copy.deepcopy(observables_card)
     oo["observables"]["XSHERANC"] = curobs
     yads = []
     p = lhapdf.mkPDF("NNPDF40_nnlo_as_01180", 0)
     # run yadism
     for kDISbThr in kDISbThrs:
-        tt = copy.deepcopy(t)
+        tt = copy.deepcopy(theory_card)
         # due to https://github.com/NNPDF/yadism/issues/167 we have to hack kqThr
         tt["kbThr"] = kDISbThr
         out = yadism.run_yadism(tt, oo)
         yad_data = out.apply_pdf_alphas_alphaqed_xir_xif(
-            p, p.alphasQ, lambda _muR: t["alphaqed"], 1.0, 1.0
+            p, p.alphasQ, lambda _muR: tt["alphaqed"], 1.0, 1.0
         )
         # yad_data = out.apply_pdf(p)
         yads.append(np.array([esf["result"] for esf in yad_data["XSHERANC"]]))
@@ -115,14 +53,6 @@ def plot_q2(x: float, q2s: npt.ArrayLike, yads: list, kDISbThrs: list):
     ax1.legend()
     fig.savefig(f"kDISbThr-x_{x}.pdf")
     plt.close(fig)
-
-
-def build_q2_obs(x: float, q2s: npt.ArrayLike, s: float = 318.0**2) -> list:
-    """Generate ESF with fixed x and varying Q2"""
-    obs = []
-    for q2 in q2s:
-        obs.append(dict(Q2=q2, x=x, y=float(q2) / s / x))
-    return obs
 
 
 # doit
